objfx fn_80098B18: t=(u8)type keeps clrlwi'd switch value in saved reg (96.71->97.49)
Target masks the int 'type' param to u8 once (clrlwi r31,r27,24) and
holds it in a saved reg across all switch compares; current used the raw
int param. 't = (u8)type' routes the value through a masked local so MWCC
keeps the clrlwi pair live in a saved reg (#83c inverse / #126). ndiff
regions 90->31. Unit 98.47->98.60. Residual is r26/r27 within-class GPR
perm + extra saved-reg numbering (#108).

(cherry picked from commit 67df499aaa78d4430c6a2ab6499807227f65b455)

+++ src/main/dll/objfx.c
@@ -1326,7 +1326,7 @@ void fn_80098B18(void* obj, f32 scale, int type, int count, int mode, f32* vec)
         params.vec[2] = z;
     }
 
-    t = type;
+    t = (u8)type;
     switch (t)
     {
     case 3:
